add more schema
